requirements.txt
streamlit==1.43.0
langchain_openai==0.3.7
python-dotenv==1.0.1